+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This parameter enables user to decide whether or not filter should be validated.
 - Added `MediatoolToDF` task class
 - Added `MediatoolToADLS` flow class
 - Added option to disable `check_dtypes_sort` in class/flow_name.
+- Added `query` parameter to `BigQueryToADLS` flow and `BigqueryToDF` task to be able to enter custom SQL query.
 
 
 ## [0.4.12] - 2023-01-31