+++ CHANGELOG.md
@@ -5,16 +5,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.4.13] - 2023-03-15
 ### Added
 - Added `validate_date_filter` parameter to `Epicor` source, `EpicorOrdersToDF` task and `EpicorOrdersToDuckDB` flow.
 This parameter enables user to decide whether or not filter should be validated.
 - Added `Mediatool` source class
 - Added `MediatoolToDF` task class
 - Added `MediatoolToADLS` flow class
-- Added option to disable `check_dtypes_sort` in class/flow_name.
+- Added option to disable `check_dtypes_sort` in `ADLSToAzureSQL` flow.
 - Added `query` parameter to `BigQueryToADLS` flow and `BigqueryToDF` task to be able to enter custom SQL query.
 - Added new end point `conversations/details/query` connection to `Genesys` task.
 - Added new task `filter_userid` in `GenesysToADLS` flow to filter out by user Ids list, previously passed by the user.
+
+### Changed
+- Changed parameter name in `BigQueryToADLS` flow - from `credentials_secret` to `credentials_key`
+
 
 ## [0.4.12] - 2023-01-31
 ### Added